factor out SMT interface
